Move generate and send incremental export to models
